Crate rrr

Structs

Beat
Stores all of the note rows that represent a beat.

BpmChange
Represents a BPM change in a chart.

Chart
A space/memory efficient representation of a chart.

CompiledChart
A runtime efficient representation of a chart used by an RRR instance.

Note
Represents a single note in a note row.

NoteRow
Stores the notes that belong to a single row in a beat.

RRR

Settings
Stores RRR settings to play charts with.

SettingsBuilder
A builder for a Settings instance.